ai_part/ocular_diagnosis_image_ai_system/requirements.txt
tensorflow==2.19.0
pandas==2.0.3
scikit-learn==1.3.0
pyyaml==6.0
matplotlib==3.7.1
seaborn==0.12.2
kaggle==1.5.16
openpyxl==3.1.2 
tqdm>=4.65.0
pytest>=7.4.0